refactor out SymbolRefSet into grammar module
